refactor(core): extract ExtensionAppLauncher and fix background launch dispatch
- extract openExtensionApp into dedicated ExtensionAppLauncher
- keep app metadata in EXTENSION_APPS
- use direct handler invocation when running in background
- keep runtime messaging for UI contexts
- fail fast when background handler is unavailable
- remove launcher logic from helpers
