+++ src/apps/options/OptionsSidebar.vue
@@ -100,7 +100,7 @@ import { useUnifiedI18n } from '@/composables/shared/useUnifiedI18n.js'
 import { getScopedLogger } from '@/shared/logging/logger.js'
 import { LOG_COMPONENTS } from '@/shared/logging/logConstants.js'
 import { REPO_URLS } from '@/shared/constants/core.js'
-import { openExtensionApp } from '@/core/helpers.js'
+import { openExtensionApp } from '@/core/ExtensionAppLauncher.js'
 import ThemeSelector from './components/ThemeSelector.vue'
 import InterfaceLocaleSelector from './components/InterfaceLocaleSelector.vue'
